Had to create a new table_utils module due to ImportError problems that I don't understand
