+++ halotools/utils/__init__.py
@@ -10,3 +10,4 @@ from .match import *
 from .spherical_geometry import *
 from .array_utils import *
 from .io_utils import *
+from .table_utils import *
